+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * image upload (https://github.com/dadi/cdn/issues/130). Allows configuration of CDN to accept image uploads. See documentation at http://docs.dadi.tech/cdn/concepts/upload
 * external image requests now possible (https://github.com/dadi/cdn/issues/151). See documentation at http://docs.dadi.tech/cdn/
+* validation added to route and recipe names, to ensure they are more than 5 characters
 * CDN can be configured to respond to the route `/robots.txt` (https://github.com/dadi/cdn/issues/153). Specify the path to a robots.txt file in the configuration file:
 
 ```json